# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Go + Gin backend for a multivendor eCommerce platform. Auth, vendor onboarding (KYC), product catalog, order splitting, commission, settlement.",
        "title": "Coolmate Multivendor eCommerce API",
        "termsOfService": "https://example.com/terms",
        "contact": {
            "name": "Coolmate Team",
            "email": "dev@coolmate.local"
        },
        "license": {
            "name": "MIT"
        },
        "version": "0.1"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/admin/orders": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders (Admin)"
                ],
                "summary": "List all orders (STUB, admin)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_coolmate_ecommerce-backend_internal_utils.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/customer/cart": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cart \u0026 Orders (Customer)"
                ],
                "summary": "Get current cart (STUB)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_coolmate_ecommerce-backend_internal_utils.APIResponse"
                        }
                    }
                }
            }
        },
        "/customer/cart/apply-coupon": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cart \u0026 Orders (Customer)"
                ],
                "summary": "Apply coupon code to cart (STUB)",
                "parameters": [
                    {
                        "description": "{ code }",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_coolmate_ecommerce-backend_internal_utils.APIResponse"
                        }
                    }
                }
            }
        },
        "/customer/cart/items": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cart \u0026 Orders (Customer)"
                ],
                "summary": "Add item to cart (STUB)",
                "parameters": [
                    {
                        "description": "{ productId, quantity }",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_coolmate_ecommerce-backend_internal_utils.APIResponse"
                        }
                    }
                }
            }
        },
        "/customer/cart/items/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cart \u0026 Orders (Customer)"
                ],
                "summary": "Update cart item quantity (STUB)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Cart item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_coolmate_ecommerce-backend_internal_utils.APIResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cart \u0026 Orders (Customer)"
                ],
                "summary": "Remove item from cart (STUB)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Cart item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_coolmate_ecommerce-backend_internal_utils.APIResponse"
                        }
                    }
                }
            }
        },
        "/customer/orders": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cart \u0026 Orders (Customer)"
                ],
                "summary": "List customer order history (STUB)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_coolmate_ecommerce-backend_internal_utils.APIResponse"
                        }
                    }
                }
            }
        },
        "/customer/orders/checkout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Should split master order into per-vendor sub-orders, apply commission, deduct stock. Currently returns 201 without side-effects.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cart \u0026 Orders (Customer)"
                ],
                "summary": "Place order — checkout cart (STUB — MVP blocker per audit)",
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_coolmate_ecommerce-backend_internal_utils.APIResponse"
                        }
                    }
                }
            }
        },
        "/customer/orders/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cart \u0026 Orders (Customer)"
                ],
                "summary": "Get order detail (STUB)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_coolmate_ecommerce-backend_internal_utils.APIResponse"
                        }
                    }
                }
            }
        },
        "/customer/returns": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Returns (Customer)"
                ],
                "summary": "List customer returns (STUB)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_coolmate_ecommerce-backend_internal_utils.APIResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Returns (Customer)"
                ],
                "summary": "Initiate return request (STUB)",
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_coolmate_ecommerce-backend_internal_utils.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/vendor/orders": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders (Vendor)"
                ],
                "summary": "List vendor sub-orders (STUB)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_coolmate_ecommerce-backend_internal_utils.APIResponse"
                        }
                    }
                }
            }
        },
        "/vendor/orders/{id}/status": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Vendor can only advance Pending → Ready to Ship (SRS §7.2). Not enforced yet.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders (Vendor)"
                ],
                "summary": "Update sub-order status (STUB)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Sub-order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "{ status }",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_coolmate_ecommerce-backend_internal_utils.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_coolmate_ecommerce-backend_internal_utils.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Paste: `Bearer \u003caccessToken\u003e`",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}