# internal/handlers/order_handler.go
package handlers

import (
	"net/http"

	"github.com/gin-gonic/gin"
	"github.com/coolmate/ecommerce-backend/internal/utils"
)

type OrderHandler struct {
	// orderService *services.OrderService
}

func NewOrderHandler(orderService interface{}) *OrderHandler {
	return &OrderHandler{}
}

// GetCart godoc
// @Summary      Get current cart (STUB)
// @Tags         Cart & Orders (Customer)
// @Security     BearerAuth
// @Produce      json
// @Success      200  {object}  utils.APIResponse
// @Router       /customer/cart [get]
func (oh *OrderHandler) GetCart(c *gin.Context) {
	utils.SuccessResponse(c, http.StatusOK, "Get cart", nil)
}

// AddToCart godoc
// @Summary      Add item to cart (STUB)
// @Tags         Cart & Orders (Customer)
// @Security     BearerAuth
// @Accept       json
// @Produce      json
// @Param        body  body      map[string]interface{}  true  "{ productId, quantity }"
// @Success      200   {object}  utils.APIResponse
// @Router       /customer/cart/items [post]
func (oh *OrderHandler) AddToCart(c *gin.Context) {
	utils.SuccessResponse(c, http.StatusOK, "Add to cart", nil)
}

// UpdateCartItem godoc
// @Summary      Update cart item quantity (STUB)
// @Tags         Cart & Orders (Customer)
// @Security     BearerAuth
// @Produce      json
// @Param        id   path      int  true  "Cart item ID"
// @Success      200  {object}  utils.APIResponse
// @Router       /customer/cart/items/{id} [put]
func (oh *OrderHandler) UpdateCartItem(c *gin.Context) {
	utils.SuccessResponse(c, http.StatusOK, "Update cart item", nil)
}

// RemoveFromCart godoc
// @Summary      Remove item from cart (STUB)
// @Tags         Cart & Orders (Customer)
// @Security     BearerAuth
// @Produce      json
// @Param        id   path      int  true  "Cart item ID"
// @Success      200  {object}  utils.APIResponse
// @Router       /customer/cart/items/{id} [delete]
func (oh *OrderHandler) RemoveFromCart(c *gin.Context) {
	utils.SuccessResponse(c, http.StatusOK, "Remove from cart", nil)
}

// ApplyCoupon godoc
// @Summary      Apply coupon code to cart (STUB)
// @Tags         Cart & Orders (Customer)
// @Security     BearerAuth
// @Accept       json
// @Produce      json
// @Param        body  body      map[string]string  true  "{ code }"
// @Success      200   {object}  utils.APIResponse
// @Router       /customer/cart/apply-coupon [post]
func (oh *OrderHandler) ApplyCoupon(c *gin.Context) {
	utils.SuccessResponse(c, http.StatusOK, "Apply coupon", nil)
}

// Checkout godoc
// @Summary      Place order — checkout cart (STUB — MVP blocker per audit)
// @Description  Should split master order into per-vendor sub-orders, apply commission, deduct stock. Currently returns 201 without side-effects.
// @Tags         Cart & Orders (Customer)
// @Security     BearerAuth
// @Accept       json
// @Produce      json
// @Success      201  {object}  utils.APIResponse
// @Router       /customer/orders/checkout [post]
func (oh *OrderHandler) Checkout(c *gin.Context) {
	utils.SuccessResponse(c, http.StatusCreated, "Checkout", nil)
}

// ListCustomerOrders godoc
// @Summary      List customer order history (STUB)
// @Tags         Cart & Orders (Customer)
// @Security     BearerAuth
// @Produce      json
// @Success      200  {object}  utils.APIResponse
// @Router       /customer/orders [get]
func (oh *OrderHandler) ListCustomerOrders(c *gin.Context) {
	utils.SuccessResponse(c, http.StatusOK, "List customer orders", nil)
}

// GetOrderDetail godoc
// @Summary      Get order detail (STUB)
// @Tags         Cart & Orders (Customer)
// @Security     BearerAuth
// @Produce      json
// @Param        id   path      int  true  "Order ID"
// @Success      200  {object}  utils.APIResponse
// @Router       /customer/orders/{id} [get]
func (oh *OrderHandler) GetOrderDetail(c *gin.Context) {
	utils.SuccessResponse(c, http.StatusOK, "Get order detail", nil)
}

// ListVendorOrders godoc
// @Summary      List vendor sub-orders (STUB)
// @Tags         Orders (Vendor)
// @Security     BearerAuth
// @Produce      json
// @Success      200  {object}  utils.APIResponse
// @Router       /vendor/orders [get]
func (oh *OrderHandler) ListVendorOrders(c *gin.Context) {
	utils.SuccessResponse(c, http.StatusOK, "List vendor orders", nil)
}

// UpdateOrderStatus godoc
// @Summary      Update sub-order status (STUB)
// @Description  Vendor can only advance Pending → Ready to Ship (SRS §7.2). Not enforced yet.
// @Tags         Orders (Vendor)
// @Security     BearerAuth
// @Accept       json
// @Produce      json
// @Param        id    path      int                true  "Sub-order ID"
// @Param        body  body      map[string]string  true  "{ status }"
// @Success      200   {object}  utils.APIResponse
// @Router       /vendor/orders/{id}/status [put]
func (oh *OrderHandler) UpdateOrderStatus(c *gin.Context) {
	utils.SuccessResponse(c, http.StatusOK, "Update order status", nil)
}

// ListAllOrders godoc
// @Summary      List all orders (STUB, admin)
// @Tags         Orders (Admin)
// @Security     BearerAuth
// @Produce      json
// @Success      200  {object}  utils.APIResponse
// @Router       /admin/orders [get]
func (oh *OrderHandler) ListAllOrders(c *gin.Context) {
	utils.SuccessResponse(c, http.StatusOK, "List all orders", nil)
}

// InitiateReturn godoc
// @Summary      Initiate return request (STUB)
// @Tags         Returns (Customer)
// @Security     BearerAuth
// @Accept       json
// @Produce      json
// @Success      201  {object}  utils.APIResponse
// @Router       /customer/returns [post]
func (oh *OrderHandler) InitiateReturn(c *gin.Context) {
	utils.SuccessResponse(c, http.StatusCreated, "Initiate return", nil)
}

// ListReturns godoc
// @Summary      List customer returns (STUB)
// @Tags         Returns (Customer)
// @Security     BearerAuth
// @Produce      json
// @Success      200  {object}  utils.APIResponse
// @Router       /customer/returns [get]
func (oh *OrderHandler) ListReturns(c *gin.Context) {
	utils.SuccessResponse(c, http.StatusOK, "List returns", nil)
}